style: move mapWithneighbours
